ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
openai==1.12.0
python-dotenv==1.0.1
httpx==0.27.0
